requirements.txt
absl-py==2.3.0
addict==2.4.0
annotated-types==0.7.0
astroid==3.3.10
asttokens==3.0.0
attrs==25.3.0
black==25.1.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7.1
contourpy==1.3.0
cycler==0.12.1
dash==3.0.4
decorator==5.2.1
dill==0.4.0
einops==0.8.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.73.0
hf-xet==1.1.4
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipython==8.18.1
ipywidgets==8.1.7
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
# Editable install with no version control (MultiScaleDeformableAttention==1.0)
-e /home/rcasal/miniconda3/envs/dac/lib/python3.9/site-packages/MultiScaleDeformableAttention-1.0-py3.9-linux-x86_64.egg
mypy==1.16.1
mypy_extensions==1.1.0
narwhals==1.44.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.18.0
onnxconverter-common==1.13.0
onnxruntime==1.19.2
onnxruntime-gpu==1.19.2
onnxruntime-tools==1.7.0
onnxsim==0.4.36
open3d==0.19.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
pluggy==1.6.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py3nvml==0.2.7
pycuda==2025.1.1
pydantic==2.11.7
pydantic_core==2.33.2
pydocstyle==6.3.0
Pygments==2.19.1
pylint==3.3.7
pyparsing==3.2.3
pyquaternion==0.9.9
pytest==8.4.0
python-dateutil==2.9.0.post0
pytools==2024.1.14
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
retrying==1.4.0
rich==14.0.0
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentry-sdk==2.30.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
snowballstemmer==3.0.1
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorrt==10.12.0.36
tensorrt_cu12==10.12.0.36
tensorrt_cu12_bindings==10.12.0.36
tensorrt_cu12_libs==10.12.0.36
threadpoolctl==3.6.0
timm==1.0.15
tomli==2.2.1
tomlkit==0.13.3
torch==2.5.1+cu124
torch_tensorrt==2.7.0
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wandb==0.20.1
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.14
xformers==0.0.28.post3
xmltodict==0.14.2
zipp==3.23.0
